requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.3.0
beautifulsoup4==4.13.5
bert-score==0.3.13
blinker==1.9.0
bs4==0.0.2
cachetools==6.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
dataclasses-json==0.6.7
distro==1.9.0
filelock==3.19.1
filetype==1.2.0
fonttools==4.60.0
frozenlist==1.7.0
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.7.0
google-api-core==2.25.1
google-auth==2.41.0
googleapis-common-protos==1.70.0
griffe==1.14.0
grpcio==1.75.1
grpcio-status==1.75.1
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.35.1
idna==3.10
Jinja2==3.1.6
jiter==0.11.0
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
langchain==0.3.27
langchain-community==0.3.30
langchain-core==0.3.76
langchain-google-genai==2.1.12
langchain-text-splitters==0.3.11
langsmith==0.4.31
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.6
mcp==1.16.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.6.4
mypy_extensions==1.1.0
narwhals==2.5.0
networkx==3.5
numpy==2.3.3
openai==1.109.1
openai-agents==0.3.3
orjson==3.11.3
packaging==25.0
pandas==2.3.2
pillow==11.3.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.32.1
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.9
pydantic-settings==2.11.0
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.1.0
rpds-py==0.27.1
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
sentence-transformers==5.1.1
setuptools==78.1.1
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.43
sse-starlette==3.0.2
starlette==0.48.0
streamlit==1.50.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.22.1
toml==0.10.2
torch==2.8.0
tornado==6.5.2
tqdm==4.67.1
transformers==4.56.2
typer==0.19.2
types-requests==2.32.4.20250913
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.37.0
wheel==0.45.1
yarl==1.20.1
zstandard==0.25.0
